removed did and replaced with actual DB assigned id

+++ mockingbird/src/main/kotlin/com/leftindust/mediq/dao/PatientDao.kt
@@ -50,7 +50,7 @@ interface PatientDao {
      */
     suspend fun removePatientByPID(pID: Int, requester: MediqToken): CustomResult<Patient, OrmFailureReason>
     suspend fun searchByName(query: String, requester: MediqToken): CustomResult<List<Patient>, OrmFailureReason>
-    suspend fun getByDoctor(did: Int, requester: MediqToken): CustomResult<List<Patient>, OrmFailureReason>
+    suspend fun getByDoctor(did: Long, requester: MediqToken): CustomResult<List<Patient>, OrmFailureReason>
     suspend fun getByVisit(vid: Long?, requester: MediqToken): CustomResult<Patient, OrmFailureReason>
     suspend fun addDoctorToPatient(
         patientInput: ID,
