+++ mockingbird/src/main/kotlin/com/leftindust/mediq/dao/VisitDao.kt
@@ -8,6 +8,6 @@ import com.leftindust.mediq.graphql.types.GraphQLVisitInput
 interface VisitDao {
     suspend fun getVisitsForPatientPid(pid: Int, requester: MediqToken): CustomResult<List<Visit>, OrmFailureReason>
     suspend fun getVisitByVid(vid: Long, requester: MediqToken): CustomResult<Visit, OrmFailureReason>
-    suspend fun getVisitsByDoctor(did: Int, requester: MediqToken): CustomResult<List<Visit>, OrmFailureReason>
+    suspend fun getVisitsByDoctor(did: Long, requester: MediqToken): CustomResult<List<Visit>, OrmFailureReason>
     suspend fun addVisit(visitInput: GraphQLVisitInput, requester: MediqToken): CustomResult<Visit, OrmFailureReason>
 }
